Add helper functions to remove deps from global sets
There are a few places where set.discard() is called instead of
set.remove(), but the intention of set.remove() is more clear.

Add helper functions for set.remove() that can be called to avoid the
try/except at call sites. Call sites will be updated in the next commit.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>