Remove unresolvable names from buildreq_cache
The buildreq cache should only contain names/symbols/etc that are
resolvable at build time. The names may indicate missing packages, but
they may also be invalid, so it is best to avoid adding the names to the
cache in the first place.

If the name is only present in the cache, retry the build. Any presence
of the name in buildreq_add will halt the build.

Fixes #480

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>